pyproject.toml
[tool.poetry]
name = "banking-fullstack"
version = "0.1.0"
description = ""
authors = ["제민욱 <minkj1992@gmail.com>"]
readme = "README.md"
packages = [{include = "django_project"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
django-test-plus = "^2.2.3"
django = "^4.2.1"
reportlab = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
